Utility: String::{lower,upper}case() without std::to{lower,upper}().
The standard function is stupid because it takes a locale into account
(which apparently involves a virtual dispatch and a mutex on every
call?!) but yet it operates on single bytes and thus is useless for
UTF-8 anyway.

Since we need this utility mostly for ASCII anyway, it makes no sense to
rely on such a heavyweight tool and an oneliner does just fine. Also
added a benchmark and according to it this is about 10x faster in Debug
and 30% faster in Release. On the other hand, the reportedly better and
faster (but horrendously crazy) C++11 STL way with std::use_facet, is
about 60% faster than calling std::tolower() in a loop. I suppose such
speedup is possible to achieve using SIMD, but even this is already
more than good enough.