Containers: rename String[View]::split{ => OnAny}WithoutEmptyParts().
A good point was made that there's no overload that would allow to split
on a character sequence (such as on `::` or `, `, etc.). For that the
best API name would be splitWithoutEmptyParts(), same as the
single-character version, but that was already used by the "split on
any" variant.

In order to fix that, splitWithoutEmptyParts() is now renamed to
splitOnAnyWithoutEmptyParts() and splitOnWhitespaceWithoutEmptyParts()
and the original APIs deprecated. Not removed because those are there
since last October already and quite some code relies on them already.
After some time I'll remove the deprecated APIs and then after some more
time I'll add the new implementation that treats the string as a
sequence and not as a bag of "any of these".