Containers: use larger allocation offset for types with larger alignment.
The allocator base offset is now exposed through an AllocationOffset
member and is always at least sizeof(std::size_t) but can be higher if
the type alignment requires it.

However only up to the default platform alignment, for overaligned types
we'll need to create a completely new allocator based on
Utility::allocateAligned().