Utility: "fix" IsIterable for Eigen 3.4.0 matrices.
The time lost is the time I will never get back. The code this issue
originates from got commited in Oct 2018, which means, in case I would
attempt to fix this upstream, I may have to wait for another 3 years
again to have a version with the fix released, and then 10 more to get
it pulled into even the oldest LTS distribution in use.

Not worth the time, no.