Add a CORRADE_MSVC_COMPATIBILITY option.
This will cover issues with MSVC2019+ that happen only when /permissive-
is not set -- most of the workarounds for MSVC 2019 were already
happening only when the standards conforming mode was not enabled, and I
expect this only get better with MSVC 2022.

The flag is impossible to detect via a preprocessor which means we can't
just disable / change certain parts of the API if the flag is missing,
thus it has to be done the old explicit way by setting it in CMake and
then remembering it inside configure.h. That way however avoids nasty
corner cases, especially when the API changes would change the ABI, so
that's nice.

Furthermore, to ensure /permissive- is always present when
CORRADE_MSVC_COMPATIBILITY is off, configure.h has a static_assert()
that fails when the flag is not present. This is the only (and not fully
robust) way to check the presence of the flag, but could be good enough.