Containers: C++17 std::string_view compatibility for String{,View}.
Had this implemented first but had to wait until Debug was able to
properly print the thing, heh. I'm also not using this new compatibility
in Debug/format() printers because it would mean including String.h as
well -- imagine that causing 33.8k preprocessed lines instead of 33.5
(on libc++), what a nightmare.