Utility: make Directory::map*() "just work" with empty files.
It works with read(), so why not make this case equivalent. The OS APIs
don't allow such operation (because they would return a null pointer),
but since we return a sized Array here, that's fine -- the user then
simply gets an empty array. But even in this case the deleter contains
the original fd, which is then closed as appropriate.