TestSuite: port Compare::*File* away from Utility::Directory.
The comparators now have a much more robust behavior in case of a read
failure, compared to just non-atomically checking if a file exists
before, and then attempting to read it but having no way to know if
it's really meant to be empty or there was some failure.

It's still half-messy as the result of readString() is converted into a
std::string, filenames are taken as std::strings and there's a lot of
other inefficiencies. Will be cleaned up once the porting to Path is
finished everywhere.