Utility: include the error code in error string printers.
So it's clearer what's up even if the localized message is completely
alien to who sees it or if it changes between OSes and thus it's overly
annoying to branch over all possibilities in the tests. This means the
mkpathNoPermission() test can be significantly shortened while also
being a lot more robust than just trying to create a dir in the root
directory, which might sometimes succeed.

On Windows this also means we can better verify PluginManager messages,
on Unix however not as there's just dlerror() and no error code.