Utility: print errno in all remaining Directory error cases as well.
It was done for some but not all, so let's be consistent. This also
means I have to Man Up and finally fix the "no permission" tests that
were just attempting to write under /root on all platforms. While that
causes errno 13 on Unix (permission denied), on Windows it's 2 (file not
found) because there's simply not a /root directory.

Oh, hm, actually some functions aren't printing any error messages at
all. Let's fix that next.