Utility: further cleanup in Directory tests.
In particular it makes ZERO SENSE to just repeat the same test data
array over and over in every function without even documenting why the
bytes are like that.

And overusing std::string or std::copy() when we absolutely don't need
to is also stupid.

Also the excessive scopes make no sense if we aren't verifying anything
after.