Utility: don't use std::string as a buffer inside Directory::read().
This code was apparently written before growable arrays were a thing,
with them it can be done much easier, and with the only copies done if
the allocation itself can't be grown. Before, however:

 - One copy from the static buffer to the string in each loop iteration
 - One copy of the read data every time the array grows (AFAICT the STL
   allocator interface has no concept of in-place allocation growing
   like with realloc(), so it gets copied always)
 - One copy for the output array creation, because zero-init is the
   default (sic!)
 - One std::copy() from the string to the output array

Hmmm and seems like I'm doing the same mistake of not using NoInit even
for seekable files. Let's look at that next.