Utility: add actual error reporting to Directory::list().
Ah well. Returning an empty list and saying nothing isn't the best way
to go about this.

Unfortunately this also uncovered a problematic behavior in
PluginManager, where a non-existent path was often used to prevent the
manager from loading dynamic plugins. Because changing that would be a
breaking change on its own, the previous behavior is preserved -- if the
directory doesn't exist, Directory::list() is not even called. On a
positive side, the new behavior will at least uncover cases where the
directory does exist but can't be accessed for whatever reason.