Utility: use asserts for failure handling in remaining Directory APIs.
The libraryLocation(), executableLocation(), home(), configurationDir()
and tmp() *do* make sense to have a failure state (theoretically when an
application would be sandboxed with no access to a filesystem), but
right now the failure states are either impossible on known systems (in
which case I added asserts) or too annoying to reproduce (such as
unsetting $HOME), in which case I added failure branches that are
however uncovered by tests.

So, like will be done for read() etc., these will all have the return
type wrapped in an Optional as well. The inverse scenario, where an
unlikely error would cause a temporary directory be placed in CWD
because the function returns an empty string in that case, isn't really
great -- it would either be silently broken, polluting user directories
with junk, or would just cause the application to crash because it
attempts to write to some readonly system location for no reason.