Containers: rename internal StringView helper functions.
There won't be any public findFirst(), only find() as that's the most
used / most common name, except for the odd one out which is
std::string::find_first_of(). And then there will be findLast(), not
rfind() or any other strange name inspired by STL.

The sudden need to add a namespace to the call is a bit silly, yes, but
what can I do.