Containers: don't set any flags in case of String::find() failure.
It just doesn't make sense, as the failed view isn't meant to be used in
any way. Realized while implementing findOr(), where attempting to
preserve the flags would lead to quite a lot of additional logic.

Also expanded the tests to check for this case, especially the "not
found" case wasn't really verified.