Containers: add explicit bool conversion to String and StringView.
Took quite a while because I couldn't figure out what the behavior
should be -- having it the same as with Array[View] where it returns
false even if the pointer is non-null would be rather stupid, as a
String is non-null always, and empty slices of non-null strings as well.
On the other hand, having it return false if the string is empty would
still mean true is returned for e.g. StringView{nullptr, 5}, which isn't
good (and can't just drop the "sized null" functionality as it's useful
in certain cases).

Fortunately I finally realized that returning true only if it's both
non-empty and non-null is probably the best option, and if users need
to check either of the case, they can query directly data() or isEmpty()
-- of course those won't work as nicely in cases like below, but
important is that the most common case finally *will*:

    if(const StringView found = string.find('@')) {
        // ...
    }

There's still the leftover TODO to (probably?) do the same change for
Array/ArrayView, but there it's blocked by MSVC, where having both a
pointer conversion and bool conversion leads to ambigous overloads. That
only gets fixed with /permissive- enabled, but that would mean dropping
MSVC 2017 support (where this flag ICEs and misbehaves), and enforcing
the flag for everybody. This is not a problem for String[View], because
it doesn't have any pointer conversion operator.