Utility: deprecate Directory in favor of a new Path namespace.
Started as an optimistic task of porting away from std::string, ended up
being a full redesign to fix other outstanding issues like a lack of
clearly distinguishable failure states for most of the APIs.

I tried to make the sweeping changes in-place at first, but it proved to
be a backwards compatibility hell that would make everyone angry. So
instead it's a whole new namespace and I took that as an opportunity to
also fix various naming issues.

The Directory namespace is now implemented in terms of the Path
functionality and isn't really any slower than it was before. On the
other hand, thanks to string views there's no useless string allocations
anymore and Path::readString() can now take over the memory ownership
coming from read(), being no longer 2x slower.

While all Directory APIs are marked as deprecated, existing code is not
updated yet, only the DirectoryTest failure message tests to make the
test still pass. Updating to the new APIs will be done in next commits.