Containers: String-to-Array should include terminating \0 in all cases.
Accidentally it wasn't allocating for it in case a SSO string was
converted to an Array, now it does. This was especially problematic with
existing code such as

    Containers::Array<char> data = Utility::format("{}", 1);

Before the returned Array contained a \0 always, but now with the return
type changed to String and the Array being subsequently implicitly
converted from it, it lose the \0 in the process, causing suffering and
misery.