Containers: make non-SSO String layout match StringView layout.
Might theoretically be useful for taking an Array<String> and turning it
into a StridedArrayView1D<StringView> without having to allocate a
temporary copy. However:

 - the ifdefs for BE will likely become a maintenance
   nightmare over time (unless BE support gets dropped)
 - since it can't work for SSO strings, benefits are questionable
   ("let's save an allocation of a large temporary array by making
   several small allocations to de-SSO the strings first")
 - and it would mean any APIs taking ArrayView<StringView> now have to
   take a StridedArrayView as well ... imagine the pain with ambiguous
   overloads

TODO: also now that we're so close it might make sense to have the
  unused bit in String size be always 1, matching the NullTerminated
  flag of StringView? But then every size retrieval gets "a bit" more
  complicated again.

Putting this aside until I'm clearer on whether this has any real use.