Utility: add String::{upper,lower}case() Containers::String overloads.
Because if the string can be updated in-place it's much more efficient
than making a copy first. Did the same with
Path::{from,to}NativeSeparators() already.

Also this way it will prevent ambigous overloads when a
Containers::String is passed to it and Containers/StringStl.h is
included -- otherwise both the StringView and std::string overloads
would be candidates.