TestSuite: remove all remaining uses of std::string.
For backwards compatibility the StringStl.h header is included, which
will provide implicit conversions to a StringView. However, it includes
just a forward declaration for std::string, so in rare cases it may
cause breakages where existing code just relied on <string> getting
included implicitly. Such case was in ResourceStaticTest, where I had to
add an explicit #include <string>.

The Tester::set*Name() APIs now take just a StringView, with no
String&& etc overloads anymore, as together with the implicit conversion
from STL it would cause a lot of ambiguous overload headaches. On the
other hand, if the input string is a global literal, it will now avoid
all copies, saving just a reference.

Well, to be frank -- of course std::stringstream is still transitively
used because Debug relies on it and std::string is also used in
Arguments, but those are the only remaining isolated cases.