Containers: delete String[Views]::except{Prefix,Suffix}(char) overloads.
Added just about a week ago in 81ebf888adb16729bf0e199c846bdd3966aafe31,
so I don't think anybody is heavily using them yet. Except for myself in
just two places that were ported to the new Utility::Path.

Because turns out, when I renamed the size-taking except() to
excepSuffix() and prefix() to exceptPrefix(), exceptPrefix(1) is now
ambiguous, unable to pick between the char and size_t overload. That's
not good and could be probably fixed to redirect integers to the size_t
overload, but perhaps even worse is that then, *in some weird cases*,
such as

    string.exceptPrefix(std::toupper('a'));

it would actually try to call exceptPrefix(65) -- because, the horror,
std::toupper() SOMEHOW takes an int and returns an int, not a char.
Thus, to avoid having yet another footgun, these overloads are now
disabled -- calling them with char literals will complain, while calling
them with int or size_t will work as expected, and with no ambiguous
overload issues.