Utility: make Resource::compile*() APIs private.
I doubt anyone ever used them directly instead of through corrade-rc,
and so they just bloat the header for no reason. Plus given that there's
an implicit expectations of any changes being backwards-compatible,
changing these APIs to implement new features such as null-terminated or
aligned data would be unnecessarily hard as I'd have to keep the old
API.

Thus, because these compilation functions are only used by corrade-rc
and its accompanying test, they're not even compiled into CorradeUtility
anymore, living just in a private header. With everything inline,
because it's not meant to be exported and it's used on a single place
(plus a test). Interestingly enough, while libCorradeUtility.so dropped
in size by 20 kB (from 490k to 462k) in Release and 200 kB in Debug
(from 3.7M to 3.5M), the corrade-rc utility got smaller by 20 kB (from
335 kB to 315 kB) in Release as well, even though it wasn't using any
actual Resource API whatsoever. I suppose such a huge change is
only because of the exported thread-local variable alone? Huh.

Additionally, the test is split into a compilation and runtime part as
well. Besides this making the STL insanities better contained and not
leaking to tests of the public API, this is an important separation in
case of a breakage -- if the code generation gets broken, there would be
no way to run the test to see what's wrong, because the test itself
wouldn't compile.

For obvious reasons, no backwards compatibility is provided. If you had
a use case for the now-removed APIs, please complain, but I don't expect
anybody had.