Containers: drop useless l-value *this restriction of String APIs.
I just tried typing

    String{}.find(' ')

and it Just Worked, so what gives? I suppose it picked the const&
overload or something, which I shouldn't have there and instead should
have a && = delete overload, typing 2x more stuff again Because It's A C
Plus Plus We're Dealing With Here.

But even if that worked, it would prevent convenient and often-used
scenarios such as

    String empty = (*Utility::Path::read("whitespace.txt")).trimmed();

and such, so that feature wouldn't have made sense anyway. Finally, all
slicing APIs inherited from Array no longer have any of that anymore
either, for the same reason, so why have it inconsistently in one place
or another.

Alternatively, there could be && overloads that return the String itself
instead of a view, but let's say what would a trimmed() do? It would
either have to mutate its own data, which isn't cheap when you're
trimming the first byte of a 500 MB JSON, or even, if the String data
were coming from nullTerminatedView(), allocating a whole new 500 MB of
JSON misery. Which is anything but good, compared to just returning a
view on (potentially dangling) itself.