Utility: port Resource away from STL.
The list() function now returns an array of string views and get() is
deprecated in favor of getString() returning a string view as well. The
returned string views have the Global flag set where possible, with null
termination being reserved for a future opt-in feature.

The internals were reworked as well, replacing a lot of ArrayViews with
nicer StringView APIs and getting rid of many unnecessary allocations --
we can now store filenames and group names as string views pointing to
the compiled-in data, so there's no need to allocate anything.

Binary-size-wise, this reduced libCorradeUtility.so by about 3 kB (461
before, 458 after) in Release. Not much, but it's going in the right
direction at least.