Utility: don't call getcwd() with a size that's not there.
The original code in Directory::current() used a std::string, so this
made sense. But for the Array it doesn't, as it has no implicit null
terminator after the size -- causing either the assert below to blow up
when the retrieved path is equal to path size, or an ASan failure, or a
crash in case things get really unlucky and we hit a page boundary.