Utility: new JSON tokenizer / parser class.
Because doing this from scratch was the easiest option after all. Yes,
jsmn is a great ... tokenizer, but when actual parsing needs to be done
as well, it quickly gets very nasty.

This has about the same memory footprint as jsmn but can also parse the
stuff (and store it in the same memory footprint). More convenience APIs
like object/array iterating and value array access and operator[] coming
next, a JSON writer that's able to consume these tokens eventually also.