Containers: document an important std::unique_ptr wart that we don't have.
The STL doesn't cease to amaze me. I wouldn't even guess the
std::unique_ptr was not const-correct until I saw the
std::indirect_value proposal and thought "wow why even bother with a new
type". What was the point of having such behavior on an *owning*
container? Or did they just make it so because the language allowed
that?!