[wip] Utility: gather resource dependencies by executing corrade-rc.
And not by reimplementing the conf file parsing with CMake. This will
allow for much easier modifications to the resource file format, such as
shorthand lists or directory globbing.

TODO: but of course it would work only in projects using Corrade and not
  in Corrade itself! Because guess what, at the time I need to gather
  dependencies, I don't yet have the corrade-rc executable built. So I
  can't use it in execute_process(), but only later in
  add_custom_command(), which means it can't know the dependency tree
  until it actually starts building stuff and that's quite shit.
TODO: there's a DEPFILE option which could solve this, but its support
  is sparse and mostly just in newer CMakes. So a no-go.
TODO: or I would need to keep the nasty regex in there for when using
  corrade_add_resource() right inside Corrade, and switch to the clean
  option only in dependent projects, but that doesn't fix the problem of
  having to maintain two different parsers for the same thing.
TODO: this also needs a changelog entry