Utility: maintain and expose current array size in JsonWriter.
It's a useful functionality and while doable on user side, it's much
easier and less error-prone to do it here. It also simplifies the
internals quite a bit.

To future- and bloat-proof, I made the array size 64bit on 64bit
systems. And because it's stored in a pair with the indentation level,
that one now has to be 64bit as well, due to padding. Hopefully won't be
too much of a problem, worst case I could pack both into a single 64bit
value, using e.g. 16 bits for indentation and 48 for size.