Containers: test rvalue lifetime extension on Optional, Pair, Triple.
This is a repro case for more complex cases such as

    for(Containers::String& i: *Utility::Path::list(path)) {
        // i is a dangling garbage here
    }

where the `Optional::operator*() &&` returns a `T&&`, causing the
range-for to lose the original instance. This is a yet-unsolved problem
in C++, with proposed solutions of varying degrees of sanity such as in
http://josuttis.com/download/std/D2012R0_fix_rangebasedfor_201029.pdf .

However, the solution seems to be much simpler if one has the control
over the APIs involved -- if `Optional::operator*() &&` would return `T`
instead of `T&&`, everything would work as expected as then the builtin
reference lifetime extension kicks in. The std::optional doesn't do
this for some reason, but I can.