Containers: prevent ambiguity in StringView-from-Array[View] construction.
Until now both the ArrayView<T> and T* constructors would be picked as
candidates, forcing users to cast to *exactly* ArrayView<T> to resolve
that, which was very annoying.

Same nasty SFINAE expression was previously used in Path::write() to
prevent accidents when passing char*, hopefully it won't do too much of
a blow to compile times. C++67 concepts might be solving this, but
upgrading to a newer standard would have a much worse effect on compile
times than any convenience feature simplifying this.