Utility: support signed 64-bit ints in Json on 32-bit.
Literally I was just missing *the last bit* to make it work. It was all
there, I just didn't see it, only wondered what *the last bit*, the sign
bit, could be abused for.

Doubles are amazing, eh? It's like in 1985 they knew that 30 years later
the world would be built on top of JavaScript and we'd need such tricks
to make stuff fast and memory-efficient.