Utility: print error in Json::parse*() also on token type mismatch.
It was originally a silent null, but that only makes it more annoying on
user side because they then have to go the extra step in providing
additional info about why it failed. With this behavior, if they don't
want the builtin message, they can check the token type on their own and
make their own handling, and only then attempt to call parse*().