Utility: mark the JSON stuff as experimental.
I always think I made the *last* breaking change and then the next idea
for a more convenient API comes. So better mark this as experimental to
tell let people to not rely too much on these yet -- not just API
breakages, but also the overall workflow may still change.