Utility: clarify allocation and lookup complexity in the Json class.
About the O(n) lookup, reduced the strong language a bit, as the snippet
in question is actually *also* O(n^2), so not really better in terms of
asymptotic complexity.

Plus, (shower thought), I think I finally have an idea how to make an
opt-in O(1) lookup for objects and arrays, by allocating a shadow lookup
table that'll be referenced from the object/array token "child count"
field. I.e., if a particular option would be enabled, a pointer gets
stored instead of child count, and child count then stored externally
together with either token offsets for array items or some kind of a
flat hashmap. This also nicely builds upon the recently added opt-in
parsing of arrays and objects, where a hashmap could be generated only
on-demand and not implicitly for all.

Originally I played with ideas like reordering the token array *somehow*
to have a O(log n) lookup at least, but since the tokens can have any
number of children, it wouldn't really be possible to perform a binary
search there. Or, hm, maybe now if I built the token offsets the same
way as for arrays and then sorted the list of offset by name, I could do
a binary search on that. Yeah great. Sorry to whoever is reading this
commit message, it's getting way too long already. Anyway, I bet that I
forget I wrote all this there and would have to reinvent this feature
from scratch when I get to actually implementing it.

This is very similar to how external strings are referenced, including
how their storage is pre-allocated to have stable pointers. I "just"
need a nice flat hashmap to use here.