Utility: make it possible to supply filename also to Json::fromString().
Because the major use case I have is to call Json::fromString() in a
glTF parser, passing an *owned* string to it -- i.e., it would make no
sense to call Json::fromFile() and have it allocate another copy just to
have nicer error messages.

The initial line/column offset is a niche use case in general, but if
I'm doing this already, why not also have semi-useful error reporting
for *.glb files as well, where the JSON is prepended with a bunch of
binary data.