Utility: make it possible to print std::filesystem::path with Debug.
This type is very special because it has a begin() / end() that return
std::filesystem::path *again*, so Debug helpfully assumes it's a nested
iterable container and then dies on infinite recursion.

Instead, it has to be treated as string-like, but then there's an
ambiguity between an implicit conversion to std::string and a builtin
iostream operator<<. For that the operator<<(Debug&, const std::string&)
had to get changed to not get selected if a type convertible to a
std::string has an iostream operator<< as well -- the assumption is that
using the operator<<() is cheaper since it doesn't require allocating a
string copy.