Utility: make Path::move() use MoveFileEx() on Windows.
This should prevent failures when the destination already exists. Also
added a test to check the behaviour.

Unfortunately ths also means that this function joins the group of Path
APIs not implemented on UWP / RT. Not sure what is one supposed to use
there, but Win32 APIs aren't that, apparently. Which makes me wonder if
the original _wrename() even worked there.

Co-authored-by: Vladimír Vondruš <mosra@centrum.cz>