PluginManager: add a Manager::externalManager<T>() getter.
So e.g. the new Magnum GltfSceneConverter can instantiate arbitrary
ImageConverter plugins instead of having to rely on a few hardcoded
formats. It was already possible to do this with the same plugin manager
interface and it's how AnyImageImporter and such work, but now it's
possible also across different plugin interfaces.

Since this makes registerExternalManager() useful in case of platforms
that have only static plugins, the internals that used to be enabled
only for platforms with dynamic plugins are now enabled everywhere.
Including the assertion that checked proper destruction order --
otherwise a plugin that retrieved the externalManager<T>() could end up
with a dangling pointer.