Containers: switch to debug asserts in StridedArrayView access APIs.
And for simple slicing, since prefix() / suffix() etc are used quite
often. On the other hand, complex slicing to members, broadcasting and
other operations aren't usually called in tight loops (if they would,
they could be done outside that loop), so those stay.

And constructors stay as well, as -- unlike slicing and element access
-- those are a common source of errors, and again it's very uncommon
that something as complex as a strided array view would need to be
formed inside a tight loop.