Rethink the CORRADE_TARGET_NEON* defines.
Turns out they're for something slightly different than I expected, in
particular conversion between 16-bit and 32-bit floats being in basic
NEON already, not some extension on top. Which makes sense, and so we
don't need a define for it. A separate problem is that the half-floats
might be ARM's own variant instead of the IEEE standard, but as that
*also* doesn't seem to be a feature available only on certain
processors, it doesn't need a define either.

Given that half-float conversions are basic NEON, the
CORRADE_TARGET_NEON_FMA then is a superse of just CORRADE_TARGET_NEON.
It also wasn't checked properly on ARM64 GCC, only on the Android ARM64
Clang, so I fixed that.

Then there's actual ARMv8.2-a half-float vector arithmetic, which I
don't really need at the moment, but since we have
CORRADE_TARGET_NEON_FP16 already, why not keep it but fix it to actually
check this feature. It also has FMA builtin, so no need for another
CORRADE_TARGET_NEON_FP16_FMA macro controlling just that.

The documentation is really bad for the whole ARM ecosystem, with only
random PDFs you can't link to, and basically nonexistent Wikipedia
presence, so I hope I didn't mess up yet again.