Rethink how CORRADE_TARGET_AVX_F16C and FMA fit into the hierarchy.
In short, they don't. There are various Atom, Celeron, VIA and old Intel
Mac chips where it's anything but AVX -> F16C -> FMA -> AVX2. There are
some processors with AVX+F16C, some with AVX+FMA, some with AVX+FMA+F16C
but not AVX2 and *also* VIA chips with AVX2 but no FMA, as noted in a
comment on https://stackoverflow.com/a/50829580.

So it's absolutely irresponsible to try to put any order into these.
There's none, and anything that wanted to rely on any such order, is
broken by design.