Cpu: disable IFUNC on GCC 4.8 if --coverage is used.
The offending flag is -fprofile-arcs and it's failing only on Ubuntu
20.04 with GCC 4.8, not with GCC 5 and not with GCC 4.8 on Arch. I
couldn't find any even vaguely related bugreport, so I suspect this has
something to with binutils or something else on that system. But since
GCC 5 works there, I don't see that much of a problem.

Unfortunately I also couldn't find any way to detect if such flag is
present, the predefined variables are exactly the same in both cases.
Which means this won't be able to detect if Corrade is built without
--coverage and then used in code that has --coverage enabled. But as I'm
probably the only person using --converage on GCC 4.8 (on the CI), I
don't think that's a problem either.