Cpu: inline the CPUID query in the header and make it inline assembly.
Needed for IFUNC resolvers as otherwise Cpu::runtimeFeatures() could be
a cross-SO call that crashes on Clang and older GCC. Everything seemed
to work in the localized case of CpuTest alone, but as soon as I started
using IFUNCs in other libraries, only GCC 11 continued to work and GCC
4.8 and Clang started to crashing inside _dl_somewhere(). Took me a
while to discover also, due to *all the other* issues that I
encountered.

Because it's essential for the header to be small, it doesn't drag in
37kLOC of <immintrin.h> (!!!) headers and instead uses inline asssembly.

Looking at the diff, should have done that right from the start instead
of cursing at pointless compiler differences. Six lines less, with no
functional change, and no header dependencies.