Cpu: give up on extra instruction sets on GCC 4.8.
I was happy when I realized that if I enable a single target and then
include an intrinsics header, the instruction set is then not tied to
other targets enabled at the same time. I.e., if I do

    #pragma GCC push_options
    #pragma GCC target("avx")
    #include <immintrin.h>
    #pragma GCC pop_options

    #pragma GCC push_options
    #pragma GCC target("lzcnt")
    #include <lzcntintrin.h>
    #pragma GCC pop_options

(plus the __AVX__ and __LZCNT__ macro push/pop of course) instead of

    #pragma GCC push_options
    #pragma GCC target("avx")
    #pragma GCC target("lzcnt")
    #include <immintrin.h> // also grabs LZCNT
    #pragma GCC pop_options

then I can use LZCNT without having to enable AVX as well. What I didn't
know back then was that on GCC 4.8 this works the other way as well --
so now I cannot use LZCNT together with AVX, because the compiler
somehow remembers that LZCNT can be used only if target is EXACTLY what
was used at the point when the intrinsics header was included, instead
of if it just CONTAINS that. Haha. Which makes it pretty useless in
practice, since e.g. the String additions I have use LZCNT / TZCNT with
either SSE2 or AVX2. Since SSE2 is implicit, that worked, but as soon I
got to the AVX code path, it no longer did, failing with a linker error.

The commit that makes all this unnecessary on GCC 4.9 is mostly just
header updates, but also two minor code changes, and those changes seem
to be *essential* for this to work. Meaning, I just have no way to make
my use case work with GCC 4.8:
https://github.com/gcc-mirror/gcc/commit/97db2bf7fb10e7eb2e8224e0471b56976f133843

While I could work around this with nasty ifdefs in user code, I feel
like the more robust way is to just not have the CORRADE_ENABLE_ macros
for the extra instruction sets there at all, thus avoiding any further
suffering right from the start. For the "base" instruction sets it seems
fine so far, most probably because there's some inheritance / dependency
between them. I hope I wouldn't need to drop those too.