TestSuite: add Compare::Aligned and NotAligned.
Divisible and NotDivisible was too annoying to use with pointers. First
I thought I'd just add an overload that is able to accept pointer values
as well, but then realized it would be better to provide also a
dedicated failure message and print the pointer in hex. So a dedicated
type it is.