Define CORRADE_TARGET_SIMD128 only on Clang 13+ and Emscripten 2.0.18+.
While the earlier versions support a certain WIP variant of the
intrinsics, it's not all, and some of these were renamed, opcodes
renumbered etc. Doesn't really make sense to pretend they're usable in
that case.

And then emsdk DOES NOT MAKE IT ANY SIMPLER by bundling a prerelease of
it, thus we *also* have to check emscripten version to be really sure
that SIMD128 can be used. UGH