acme: unconditionally ignore #cmakedefine.
For Cpu.h as a single-header library I need to pull in most of
configure.h but without the stuff that got defined for my local build.
Copying the whole thing makes no sense from a maintainability
perspective, so instead I'm going to pull in configure.h.cmake and then
ignore the actual configure.h when included.