Containers: re-fix Iterable ambiguity w/ StridedArrayView and ArrayView.
Instead of std::is_convertible / std::is_constructible, which generally
requires the type in question to be defined (and thus causes immense
problems when Iterable is used in overload sets that take
forward-declared types), I'm not using the "priority tag" trick like
earlier in the Cpu library.

I miss the days where I could just write C++ code and it worked, without
having to pile on nasty SFINAE crap and without having to explain that
crap in lengthy comments. Sigh.